Moved the MySQL Point type into the schema manager test namespace
